\name{bimetsDataset}
\alias{BIMETS Datasets}
\alias{bimets_12F_YP2D__}
\alias{bimets_12L_YP2D__}
\alias{bimets_12_D2YP__}
\alias{bimets_1F_YP2D__}
\alias{bimets_1L_YP2D__}
\alias{bimets_1_D2YP__}
\alias{bimets_24F_YP2D__}
\alias{bimets_24L_YP2D__}
\alias{bimets_24_D2YP__}
\alias{bimets_2F_YP2D__}
\alias{bimets_2L_YP2D__}
\alias{bimets_2_D2YP__}
\alias{bimets_366_D2YP__}
\alias{bimets_366_YP2D__}
\alias{bimets_36F_YP2D__}
\alias{bimets_36L_YP2D__}
\alias{bimets_36_D2YP__}
\alias{bimets_3F_YP2D__}
\alias{bimets_3L_YP2D__}
\alias{bimets_3_D2YP__}
\alias{bimets_4F_YP2D__}
\alias{bimets_4L_YP2D__}
\alias{bimets_4_D2YP__}
\alias{bimets_53F_YP2D__}
\alias{bimets_53L_YP2D__}
\alias{bimets_53_D2YP__}
\alias{bimets_static_TD90_1__}
\alias{bimets_static_TD90_2__}
\alias{bimets_static_TD90_3__}
\alias{bimets_static_TD90_4__}
\alias{bimets_static_TD90_5__}
\alias{bimets_static_TD90_6__}
\alias{bimets_static_TD90__}
\alias{bimets_static_G90_1__}
\alias{bimets_static_G90_2__}
\alias{bimets_static_G90_3__}
\alias{bimets_static_G90_4__}
\alias{bimets_static_G90_5__}
\alias{bimets_static_G90_6__}
\alias{bimets_static_G90__}
\alias{bimets_static_startYear___}
\alias{bimets_static_totalLength___}


\title{
BIMETS User and Internal datasets
}
\description{
BIMETS package contains examples and hidden datasets that provide faster performance in time series analysis.
}


\seealso{
\code{\link[=TSERIES]{TIMESERIES}}\cr
\code{\link[=is.bimets]{is.bimets}}\cr
\code{\link[=as.bimets]{as.bimets}}\cr
\code{\link[=fromBIMETStoXTS]{fromBIMETStoXTS}}\cr
\code{\link[=fromBIMETStoTS]{fromBIMETStoTS}}\cr
\code{\link[=idxOver]{BIMETS indexing}}\cr
}

