\name{bimetsDataset}
\alias{BIMETS Datasets}
\alias{LONGBASE}
\alias{FRB__MODEL}
\alias{FRB__MCAP__WP__MODEL}


\title{
BIMETS Internal Datasets
}
\description{
BIMETS package contains example datasets in order to execute code in vignettes, and hidden technical datasets in order to allow a faster performance in time series analysis.
}


\seealso{
\code{\link[=TSERIES]{TIMESERIES}}\cr
\code{\link[=is.bimets]{is.bimets}}\cr
\code{\link[=as.bimets]{as.bimets}}\cr
\code{\link[=fromBIMETStoXTS]{fromBIMETStoXTS}}\cr
\code{\link[=fromBIMETStoTS]{fromBIMETStoTS}}\cr
\code{\link[=idxOver]{BIMETS indexing}}\cr
}

